# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aether Oracle API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  birthTime?: string | null;
  /** @nullable */
  birthPlace?: string | null;
  /** @nullable */
  photoPath?: string | null;
  /** @nullable */
  bedDirection?: string | null;
  /** @nullable */
  avatarType?: string | null;
  notificationsEnabled: boolean;
  createdAt: string;
}

export interface ProfileInput {
  name?: string;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  birthTime?: string | null;
  /** @nullable */
  birthPlace?: string | null;
  /** @nullable */
  photoPath?: string | null;
  /** @nullable */
  bedDirection?: string | null;
  /** @nullable */
  avatarType?: string | null;
  notificationsEnabled?: boolean;
}

export interface ArcanaDetail {
  number: number;
  name: string;
  essence: string;
  pros: string[];
  cons: string[];
  herbs: string[];
  oils: string[];
  cocktail: string;
  affirmation: string;
  health: string;
  energyChannels: string;
  nutrition: string;
}

export interface BaziPillar {
  name: string;
  heavenlyStem: string;
  earthlyBranch: string;
  element: string;
}

export interface BaziStar {
  name: string;
  description: string;
  advice: string;
  /** @nullable */
  sector?: string | null;
}

export interface BaziSummary {
  dayMaster: string;
  dayElement: string;
  dayElementMeaning: string;
  pillars: BaziPillar[];
  stars: BaziStar[];
}

export interface FengShuiInfo {
  direction: string;
  starNumber: number;
  starName: string;
  influence: string;
  recommendation: string;
  isUnfavorable: boolean;
}

export interface MatrixPoint {
  position: string;
  arcanaNumber: number;
  arcanaName: string;
  essence: string;
}

export interface PersonalMatrix {
  birthDate: string;
  points: MatrixPoint[];
}

export interface Feedback {
  id: number;
  forecastId: number;
  date: string;
  accuracy: string;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface DailyForecast {
  id: number;
  date: string;
  arcanaNumber: number;
  arcanaName: string;
  baziElement: string;
  hasWarning: boolean;
  synthesisText: string;
  matrix: ArcanaDetail;
  bazi: BaziSummary;
  fengShui?: FengShuiInfo | null;
  conflicts: string[];
  warnings: string[];
  feedback?: Feedback | null;
}

export interface FeedbackInput {
  accuracy: string;
  /** @nullable */
  comment?: string | null;
}

export interface Contact {
  id: number;
  name: string;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  deathDate?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  relationshipType?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  birthPlace?: string | null;
  /** @nullable */
  notes?: string | null;
  notificationDays: number;
  isActive: boolean;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  deathDate?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  relationshipType?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  birthPlace?: string | null;
  /** @nullable */
  notes?: string | null;
  notificationDays?: number;
}

export interface ContactUpdate {
  name?: string;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  deathDate?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  relationshipType?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  birthPlace?: string | null;
  /** @nullable */
  notes?: string | null;
  notificationDays?: number;
  isActive?: boolean;
}

export interface UpcomingBirthday {
  contactId: number;
  name: string;
  birthDate: string;
  daysUntil: number;
  /** @nullable */
  turningAge?: number | null;
}

export interface FamilyConnection {
  id: number;
  contactId1: number;
  contactId2: number;
  connectionType: string;
  createdAt: string;
}

export interface FamilyConnectionInput {
  contactId1: number;
  contactId2: number;
  connectionType: string;
}

export interface Dream {
  id: number;
  date: string;
  dreamText: string;
  interpretation: string;
  keywords: string[];
  createdAt: string;
}

export interface DreamInput {
  dreamText: string;
}

export interface Task {
  id: number;
  date: string;
  taskType: string;
  taskText: string;
  targetValue: number;
  actualValue: number;
  isCompleted: boolean;
  isDailyGoal: boolean;
}

export interface TaskInput {
  /** @nullable */
  date?: string | null;
  taskType: string;
  taskText: string;
  targetValue?: number;
  actualValue?: number;
  isDailyGoal?: boolean;
}

export interface TaskUpdate {
  actualValue?: number;
  isCompleted?: boolean;
}

export interface Travel {
  id: number;
  countryCode: string;
  countryName: string;
  visited: boolean;
  wishlist: boolean;
  /** @nullable */
  visitedDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface TravelInput {
  countryCode: string;
  countryName: string;
  visited?: boolean;
  wishlist?: boolean;
  /** @nullable */
  visitedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface TravelUpdate {
  visited?: boolean;
  wishlist?: boolean;
  /** @nullable */
  visitedDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface TravelStats {
  visitedCount: number;
  wishlistCount: number;
}

export interface Dashboard {
  profileComplete: boolean;
  /** @nullable */
  arcanaNumber?: number | null;
  /** @nullable */
  arcanaName?: string | null;
  /** @nullable */
  baziElement?: string | null;
  hasWarning?: boolean;
  /** @nullable */
  affirmation?: string | null;
  upcomingBirthdaysCount: number;
  waterProgress: number;
  waterTarget: number;
  stepsProgress: number;
  stepsTarget: number;
  recentDream?: Dream | null;
}

export type ListTasksParams = {
date?: string;
};