# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aether Oracle API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /forecast/{id}/feedback:
    post:
      operationId: submitFeedback
      tags: [forecast]
      summary: Submit feedback on a forecast
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "200":
          description: Saved feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /connections:
    get:
      operationId: listFamilyConnections
      tags: [contacts]
      summary: List family tree connections
      responses:
        "200":
          description: Connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyConnection"
    post:
      operationId: createFamilyConnection
      tags: [contacts]
      summary: Create a family tree connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyConnectionInput"
      responses:
        "201":
          description: Created connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyConnection"
  /connections/{id}:
    delete:
      operationId: deleteFamilyConnection
      tags: [contacts]
      summary: Delete a family tree connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /dreams:
    get:
      operationId: listDreams
      tags: [dreams]
      summary: List dream journal entries
      responses:
        "200":
          description: Dreams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dream"
    post:
      operationId: createDream
      tags: [dreams]
      summary: Record a dream and get an interpretation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DreamInput"
      responses:
        "201":
          description: Created dream with interpretation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dream"
  /dreams/{id}:
    delete:
      operationId: deleteDream
      tags: [dreams]
      summary: Delete a dream entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Feedback:
      type: object
      required: [id, forecastId, date, accuracy, createdAt]
      properties:
        id:
          type: integer
        forecastId:
          type: integer
        date:
          type: string
        accuracy:
          type: string
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    FeedbackInput:
      type: object
      required: [accuracy]
      properties:
        accuracy:
          type: string
        comment:
          type: ["string", "null"]
    Contact:
      type: object
      required: [id, name, notificationDays, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        birthDate:
          type: ["string", "null"]
        deathDate:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        relationshipType:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        notificationDays:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        birthDate:
          type: ["string", "null"]
        deathDate:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        relationshipType:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        notificationDays:
          type: integer
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        birthDate:
          type: ["string", "null"]
        deathDate:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        relationshipType:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        notificationDays:
          type: integer
        isActive:
          type: boolean
...
    FamilyConnection:
      type: object
      required: [id, contactId1, contactId2, connectionType, createdAt]
      properties:
        id:
          type: integer
        contactId1:
          type: integer
        contactId2:
          type: integer
        connectionType:
          type: string
        createdAt:
          type: string
    FamilyConnectionInput:
      type: object
      required: [contactId1, contactId2, connectionType]
      properties:
        contactId1:
          type: integer
        contactId2:
          type: integer
        connectionType:
          type: string
    Dream:
      type: object
      required: [id, date, dreamText, interpretation, keywords, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        dreamText:
          type: string
        interpretation:
          type: string
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
    DreamInput:
      type: object
      required: [dreamText]
      properties:
        dreamText:
          type: string
...