# lib/api-zod/src/generated/types/familyConnectionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Aether Oracle API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FamilyConnectionInput {
  contactId1: number;
  contactId2: number;
  connectionType: string;
}